requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
crispy-bootstrap5==2023.10; python_version >= '3.8'
cryptography==42.0.2
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.9; python_version >= '3.8'
django-allauth==0.60.1; python_version >= '3.7'
django-bootstrap-v5==1.0.11; python_version >= '3.6' and python_version < '4.0'
django-crispy-forms==2.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
django-widget-tweaks==1.5.0; python_version >= '3.8'
gunicorn==21.2.0; python_version >= '3.5'
idna==3.6; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python3-openid==3.2.0
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
